adjusted patch based on feedbakc
